Compartmentalize Onboarding components
